requirements.txt
# Core build tools
setuptools>=68.0.0
wheel>=0.41.0
pip>=23.0.0

# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data processing
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2

# ML and embeddings
sentence-transformers==2.2.2
torch==2.1.2
openai==1.3.70

# HTTP and environment
requests==2.31.0
python-dotenv==1.0.0 

# Database
SQLAlchemy==2.0.30
PyMySQL==1.1.1

# Security
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# RSS feeds
feedparser==6.0.10
feedparser>=6.0.0

gdown